refactor(packages-mcp): remove dead helpers, fix broken import, dedupe request prep
- Fix broken `../utils/modelCache.js` dynamic import in getPricing; repoint
  getImageModels to the real `../utils/models.js` export (image-pricing branch
  now works instead of always throwing)
- Delete unused coreUtils exports createToolDefinition and createErrorResponse
- Add shared postChatCompletion helper in coreUtils; dedupe the identical
  body-cleanup + 30s-timeout POST + 429/error mapping in generateText and
  chatCompletion (webSearch left untouched — distinct behavior)
- Extract prepareImageRequest/prepareVideoRequest in imageService; dedupe the
  verbatim validation + duration checks + buildQueryParams across the image and
  video binary/URL pairs
